refactor: LottoResultJudge 제거
