--- src/main/java/model/LottoRankJudge.java
@@ -1,26 +1,0 @@
-package model;
-
-public class LottoRankJudge {
-
-    private final LottoNumbers winningNumbers;
-    private final LottoNumber bonusNumber;
-
-    public LottoRankJudge(LottoNumbers winningNumbers, LottoNumber bonusNumber) {
-        this.winningNumbers = winningNumbers;
-        this.bonusNumber = bonusNumber;
-    }
-
-    public LottoRank judge(LottoNumbers lottoNumbers) {
-        int count = getMatchedCountAboutWinningNumbers(lottoNumbers);
-        boolean bonusMatch = isBonusMatch(lottoNumbers);
-        return LottoRank.of(count, bonusMatch);
-    }
-
-    private int getMatchedCountAboutWinningNumbers(LottoNumbers lottoNumbers) {
-        return winningNumbers.getMatchedNumberCountWith(lottoNumbers);
-    }
-
-    private boolean isBonusMatch(LottoNumbers lottoNumbers) {
-        return lottoNumbers.contains(bonusNumber);
-    }
-}
