+++ src/test/java/model/LottoMachineTest.java
@@ -3,13 +3,6 @@ package model;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import model.LottoMachine;
-import model.LottoNumber;
-import model.LottoNumbers;
-import model.LottoRank;
-import model.LottoRankJudge;
-import model.LottoResult;
-import model.Prize;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -36,7 +29,7 @@ public class LottoMachineTest {
 
     @BeforeEach
     void setUp() {
-        lottoMachine = new LottoMachine(new LottoRankJudge(WINNING_LOTTO_NUMBERS, BONUS));
+        lottoMachine = new LottoMachine(new WinningLottoNumbers(WINNING_LOTTO_NUMBERS, BONUS));
     }
 
     @Test
